+++ src/tests.rs
@@ -54,7 +54,7 @@ fn add_and_update() {
         .unwrap()
         .as_secs();
 
-    // insert at one instance (tests run in parallel, so add_post_success() may not be ready)
+    // insert an instance (tests run in parallel, so add_post_success() may not be ready)
     let mut add_response = client.post("/add")
         .body("url=https://privatebin.net")
         .header(ContentType::Form)
@@ -106,4 +106,59 @@ fn add_and_update() {
         .expect("selecting oldest check, now deleted");
     let empty: Vec::<i32> = vec![]; // need to do this, so Rust can infer the type of the empty vector
     assert_eq!(empty, oldest_check);
+}
+
+#[test]
+// incorporate add POST success test, as the delete requires it to exist first
+fn add_and_delete() {
+    use super::schema::checks::dsl::*;
+    use super::schema::instances;
+
+    let rocket = rocket();
+    let conn = super::DirectoryDbConn::get_one(&rocket).expect("database connection");
+    let client = Client::new(rocket).expect("valid rocket instance");
+    let now = SystemTime::now()
+        .duration_since(SystemTime::UNIX_EPOCH)
+        .unwrap()
+        .as_secs();
+
+    // insert an instance (tests run in parallel, so add_post_success() may not be ready)
+    let mut add_response = client.post("/add")
+        .body("url=http://zerobin-legacy.dssr.ch")
+        .header(ContentType::Form)
+        .dispatch();
+    assert_eq!(add_response.status(), Status::Ok);
+    assert!(add_response.body_string().map_or(false, |s| s.contains(&"Successfully added URL: ")));
+
+    // insert checks
+    let mut query = "INSERT INTO checks (updated, up, instance_id) VALUES (".to_string();
+    let mut instance_checks = vec![];
+    for interval in 0..super::MAX_FAILURES {
+        instance_checks.push(
+            format!("datetime({}, 'unixepoch'), 0, 1", now - (interval * super::CRON_INTERVAL))
+        );
+    }
+    write!(
+        &mut query,
+        "{})",
+        instance_checks.join("), (")
+    ).unwrap();
+    conn.execute(&query)
+        .expect("inserting test checks for instance ID 1");
+
+    let key = std::env::var("CRON_KEY").expect("environment variable CRON_KEY needs to be set");
+    let mut response = client.get(format!("/update/{}", key.replace("/", "%2F"))).dispatch();
+    assert_eq!(response.status(), Status::Ok);
+    assert!(response.body_string().map_or(false, |s| s.contains(&"removed instances that failed too many times")));
+    let deleted_check: Vec<i32> = checks.select(instance_id)
+        .filter(instance_id.eq(1))
+        .load(&*conn)
+        .expect("selecting check for instance 1, now deleted");
+    let empty: Vec::<i32> = vec![]; // need to do this, so Rust can infer the type of the empty vector
+    assert_eq!(empty, deleted_check);
+    let deleted_instance: Vec<i32> = instances::table.select(instances::id)
+        .filter(instances::id.eq(1))
+        .load(&*conn)
+        .expect("selecting instance 1, now deleted");
+    assert_eq!(empty, deleted_instance);
 }